Preserve the public main and compatibility branch histories while selecting the fully built and tested OpenMW 0.51 integration tree.

+++ apps/openmw/mwgui/inventorytabsoverlay.hpp
@@ -17,7 +17,7 @@ namespace MWGui
         InventoryTabsOverlay();
 
         int getHeight();
-        void setTab(int index);
+        void setTab(size_t index);
 
     private:
         std::vector<MyGUI::Button*> mTabs;
