+++ apps/openmw/mwworld/globals.cpp
@@ -100,7 +100,7 @@ namespace MWWorld
         }
     }
 
-    int Globals::countSavedGameRecords() const
+    size_t Globals::countSavedGameRecords() const
     {
         return mVariables.size();
     }
